Move helpers for validation tests to test_utils
