+++ metricflow/test/model/validations/test_configurable_rules.py
@@ -5,7 +5,7 @@ from dbt_semantic_interfaces.model_validator import ModelValidator
 from dbt_semantic_interfaces.objects.semantic_manifest import SemanticManifest
 from dbt_semantic_interfaces.objects.metric import MetricTypeParams, MetricInput, MetricType
 from dbt_semantic_interfaces.validations.metrics import DerivedMetricRule
-from metricflow.test.model.validations.helpers import metric_with_guaranteed_meta
+from dbt_semantic_interfaces.test_utils import metric_with_guaranteed_meta
 
 
 def test_can_configure_model_validator_rules(  # noqa: D
