+++ metricflow/test/model/validations/test_data_warehouse_tasks.py
@@ -18,7 +18,7 @@ from dbt_semantic_interfaces.objects.semantic_manifest import SemanticManifest
 from metricflow.protocols.async_sql_client import AsyncSqlClient
 from metricflow.sql.sql_bind_parameters import SqlBindParameters
 from metricflow.test.fixtures.setup_fixtures import MetricFlowTestSessionState
-from metricflow.test.model.validations.helpers import semantic_model_with_guaranteed_meta
+from dbt_semantic_interfaces.test_utils import semantic_model_with_guaranteed_meta
 from metricflow.test.plan_utils import assert_snapshot_text_equal, make_schema_replacement_function
