+++ metricflow/test/model/validations/test_dimension_const.py
@@ -11,7 +11,7 @@ from dbt_semantic_interfaces.references import DimensionReference, MeasureRefere
 from dbt_semantic_interfaces.validations.semantic_models import SemanticModelTimeDimensionWarningsRule
 from dbt_semantic_interfaces.validations.dimension_const import DimensionConsistencyRule
 from dbt_semantic_interfaces.validations.validator_helpers import ModelValidationException
-from metricflow.test.model.validations.helpers import semantic_model_with_guaranteed_meta, metric_with_guaranteed_meta
+from dbt_semantic_interfaces.test_utils import semantic_model_with_guaranteed_meta, metric_with_guaranteed_meta
 from dbt_semantic_interfaces.objects.time_granularity import TimeGranularity
