+++ metricflow/test/model/validations/test_semantic_models.py
@@ -2,7 +2,7 @@ import pytest
 
 from dbt_semantic_interfaces.objects.elements.dimension import Dimension, DimensionType, DimensionTypeParams
 from dbt_semantic_interfaces.validations.validator_helpers import ModelValidationException
-from metricflow.test.model.validations.helpers import semantic_model_with_guaranteed_meta
+from dbt_semantic_interfaces.test_utils import semantic_model_with_guaranteed_meta
 from dbt_semantic_interfaces.objects.time_granularity import TimeGranularity
